requirements.txt
youtube-transcript-api>=0.6.1
kivy>=2.2.1
kivymd>=1.1.1
plyer>=2.1.0
android-permissions>=1.1.0
pyjnius>=1.5.0
